bindings: python: use pkg-config where available to validate libgpiod version
Use pkg-config to check against LIBGPIOD_MINIMUM_VERSION when the required
Python dependencies - distutils.spawn, subprocess and packaging.version -
are available and find_executable("pkg-config") is successful.

Does not catch all cases, but produces an informative early-exit when the
needed tools are available and the system libgpiod version is too old.

Signed-off-by: Phil Howard <phil@gadgetoid.com>